# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin、Gorm 和 Quasar 的低代码管理底座",
        "title": "Sweet Admin",
        "contact": {
            "name": "南",
            "email": "maxdwy@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/sweet_admin",
    "paths": {
...
        "/admin/menu": {
            "post": {
                "description": "创建菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/button": {
            "post": {
                "description": "创建菜单按钮",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "创建菜单按钮",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuButtonCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/button/{id}": {
            "put": {
                "description": "更新菜单按钮",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "更新菜单按钮",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单按钮ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuButtonUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除菜单按钮",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "删除菜单按钮",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单按钮ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/buttons/{menuId}": {
            "get": {
                "description": "获取菜单按钮列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取菜单按钮列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "menuId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/my": {
            "get": {
                "description": "获取我的菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取我的菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/order": {
            "put": {
                "description": "批量更新菜单排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "更新菜单排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuOrderUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/query": {
            "post": {
                "description": "查询菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "查询菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Basic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/refresh-cache": {
            "post": {
                "description": "当前菜单读取为实时数据库查询，此接口用于前端统一触发菜单刷新",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "刷新菜单缓存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/user/{id}": {
            "get": {
                "description": "获取用户菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取用户菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/menu/{id}": {
            "get": {
                "description": "根据ID获取菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "根据ID获取菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "根据ID删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enum.SysMenuButtonDisplayMode": {
            "type": "string",
            "enum": [
                "auto",
                "icon",
                "text",
                "icon_text"
            ],
            "x-enum-varnames": [
                "ButtonDisplayAuto",
                "ButtonDisplayIcon",
                "ButtonDisplayText",
                "ButtonDisplayIconText"
            ]
        },
...
        "enum.SysMenuPageType": {
            "type": "string",
            "enum": [
                "directory",
                "fixed",
                "low_code"
            ],
            "x-enum-varnames": [
                "MenuPageTypeDirectory",
                "MenuPageTypeFixed",
                "MenuPageTypeLowCode"
            ]
        },
...
        "request.Basic": {
            "type": "object",
            "properties": {
                "expressions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.ExpressionGroup"
                    }
                },
                "filters": {
                    "description": "额外过滤条件（联动/级联）",
                    "type": "object",
                    "additionalProperties": {}
                },
                "include_deleted": {
                    "description": "是否查询删除数据",
                    "type": "boolean"
                },
                "menu_id": {
                    "description": "当前菜单ID（用于数据权限）",
                    "type": "integer"
                },
                "num": {
                    "type": "integer",
                    "example": 10
                },
                "order": {
                    "$ref": "#/definitions/request.Order"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "quick_query": {
                    "$ref": "#/definitions/request.QuickQuery"
                },
                "table_code": {
                    "type": "string",
                    "example": "sys_dict"
                }
            }
        },
...
        "request.MenuButtonCreateReq": {
            "type": "object",
            "required": [
                "code",
                "menu_id",
                "name",
                "position"
            ],
            "properties": {
                "after_hooks": {
                    "type": "string"
                },
                "api_path": {
                    "type": "string"
                },
                "before_hooks": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "confirm_text": {
                    "type": "string"
                },
                "disable_when": {
                    "type": "string"
                },
                "display_mode": {
                    "$ref": "#/definitions/enum.SysMenuButtonDisplayMode"
                },
                "event_action": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "http_method": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "is_button": {
                    "type": "boolean"
                },
                "is_disabled": {
                    "type": "boolean"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "memo": {
                    "type": "string"
                },
                "menu_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "params_schema": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/enum.SysMenuButtonPosition"
                },
                "sequence": {
                    "type": "integer"
                }
            }
        },
        "request.MenuButtonUpdateReq": {
            "type": "object",
            "required": [
                "code",
                "id",
                "menu_id",
                "name",
                "position"
            ],
            "properties": {
                "after_hooks": {
                    "type": "string"
                },
                "api_path": {
                    "type": "string"
                },
                "before_hooks": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "confirm_text": {
                    "type": "string"
                },
                "disable_when": {
                    "type": "string"
                },
                "display_mode": {
                    "$ref": "#/definitions/enum.SysMenuButtonDisplayMode"
                },
                "event_action": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "http_method": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_button": {
                    "type": "boolean"
                },
                "is_disabled": {
                    "type": "boolean"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "memo": {
                    "type": "string"
                },
                "menu_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "params_schema": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/enum.SysMenuButtonPosition"
                },
                "sequence": {
                    "type": "integer"
                }
            }
        },
        "request.MenuCreateReq": {
            "type": "object",
            "required": [
                "component",
                "is_hidden",
                "name",
                "path",
                "pid",
                "sequence",
                "title"
            ],
            "properties": {
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "option": {
                    "type": "string"
                },
                "page_type": {
                    "$ref": "#/definitions/enum.SysMenuPageType"
                },
                "path": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "redirect": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "table_code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "request.MenuOrderUpdateReq": {
            "type": "object",
            "required": [
                "menus"
            ],
            "properties": {
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.MenuOrderItemReq"
                    }
                }
            }
        },
        "request.MenuUpdateReq": {
            "type": "object",
            "required": [
                "component",
                "id",
                "is_hidden",
                "name",
                "path",
                "pid",
                "sequence",
                "title"
            ],
            "properties": {
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "option": {
                    "type": "string"
                },
                "page_type": {
                    "$ref": "#/definitions/enum.SysMenuPageType"
                },
                "path": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "redirect": {
                    "type": "string"
                },
                "sequence": {
                    "type": "integer"
                },
                "table_code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "request.QuickQuery": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}